# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LegalFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Cases ──────────────────────────────────────────────────────────────────
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{documentId}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{documentId}/analyze:
    post:
      operationId: analyzeDocument
      tags: [documents]
      summary: AI-analyze a document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAnalysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{documentId}/ask:
    post:
      operationId: askDocument
      tags: [documents]
      summary: Ask a question about a document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentQuestion"
      responses:
        "200":
          description: Answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAnswer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Activities ─────────────────────────────────────────────────────────────
...
components:
  schemas:
    Uuid:
      type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Client:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        description:
          type: string
        status:
          type: string
        category:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        aiSummary:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        risks:
          type: ["array", "null"]
          items:
            type: string
        questions:
          type: ["array", "null"]
          items:
            type: string
        documents:
          type: ["array", "null"]
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - description
        - status
        - createdAt
        - updatedAt
    ClientInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        description:
          type: string
          minLength: 1
        status:
          type: string
        category:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        aiSummary:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        risks:
          type: ["array", "null"]
          items:
            type: string
        questions:
          type: ["array", "null"]
          items:
            type: string
        documents:
          type: ["array", "null"]
          items:
            type: string
      required:
        - name
        - description
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        description:
          type: string
        status:
          type: string
        category:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        aiSummary:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        risks:
          type: ["array", "null"]
          items:
            type: string
        questions:
          type: ["array", "null"]
          items:
            type: string
        documents:
          type: ["array", "null"]
          items:
            type: string

    # ── Cases ──
...
    Document:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        clientId:
          type: ["string", "null"]
        caseId:
          type: ["string", "null"]
        title:
          type: string
        fileName:
          type: ["string", "null"]
        fileType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        storagePath:
          type: ["string", "null"]
        textContent:
          type: ["string", "null"]
        aiAnalysis:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - title
        - createdAt
        - updatedAt
    DocumentInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        clientId:
          type: ["string", "null"]
        caseId:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        fileType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        fileBase64:
          type: ["string", "null"]
      required:
        - title
    DocumentAnalysis:
      type: object
      properties:
        summary:
          type: string
        risks:
          type: array
          items:
            type: string
        clauses:
          type: array
          items:
            type: string
      required:
        - summary
        - risks
        - clauses
    DocumentQuestion:
      type: object
      properties:
        question:
          type: string
          minLength: 1
      required:
        - question
    DocumentAnswer:
      type: object
      properties:
        answer:
          type: string
      required:
        - answer

    # ── Activities ──
...